feat(MCTS): Change type 'int' to 'size_t' to pass hashed card IDs

+++ Extensions/RosettaTorch/Includes/MCTS/SOMCTS.hpp
@@ -157,7 +157,7 @@ class SOMCTS
                 m_board->GetCurrentPlayerStateRefView().GetActionValidGetter());
         }
 
-        int GetNumber(ActionType actionType, ActionChoices& choices) final
+        size_t GetNumber(ActionType actionType, ActionChoices& choices) final
         {
             if (actionType != ActionType::MAIN_ACTION)
             {
